creates styled-components globalstyles, moves more common styled components to individual folders, moves meeting detail table to separate statelessonent, adds key property to createmeeting invitees and questions
