--- src/App.css
@@ -1,27 +1,0 @@
-#app {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  overflow: hidden;
-  background: #8baaa0;
-  background-position: fixed;
-  color: #cfedfc;
-  margin-bottom: 2rem;
-}
-
-.scroller::-webkit-scrollbar-track {
-
-  border-radius: 3px;
-  background-color: #ffffff;
-}
-
-.scroller::-webkit-scrollbar {
-  width: 12px;
-  background-color: #fff;
-}
-
-.scroller::-webkit-scrollbar-thumb {
-  border-radius: 3px;
-  background-color: lightgrey;
-}
