+++ src/components/UpdateMeeting/index.js
@@ -4,7 +4,6 @@ import { callUpdateMeeting } from "../../actions/index";
 import styled from "styled-components";
 import { Calendar } from "primereact/calendar";
 import { InputText } from "primereact/inputtext";
-import { Checkbox } from "primereact/checkbox";
 import { ScrollPanel } from "primereact/scrollpanel";
 import { AutoComplete } from "primereact/autocomplete";
 import PrimaryButton from "../Common/PrimaryButton";
