+++ src/index.js
@@ -9,8 +9,6 @@ import { reducer } from "./reducers/index";
 import { persistStore, persistReducer } from "redux-persist";
 import storage from "redux-persist/lib/storage";
 
-import "./index.css";
-
 const logger = createLogger({
   collapsed: true
 });
