+++ src/screens/Register/index.js
@@ -1,16 +1,5 @@
 import React, { Component } from "react";
 import Register from "../../components/Register";
-import styled from "styled-components";
-
-const Main = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  @media (max-width: 768px) {
-    flex-direction: column;
-  }
-`;
 
 //this screen should return components necessary to build the landing page.
 
@@ -23,6 +12,5 @@ class ScreensRegister extends Component {
     );
   }
 }
-ScreensRegister.propTypes = {};
 
 export default ScreensRegister;
